Converted project to Maven Structure
Added pom.xml with necessary dependencies, organized project files according to Maven structure
added .gitignore to ignore the maven target directory
