--- src/View/GUI.java
@@ -1,17 +1,0 @@
-package View;
-import Model.Data;
-
-public class GUI
-{
-    Data data = new Data();
-
-    public GUI(Data data)
-    {
-        this.data = data;
-    }
-
-    public void Run()
-    {
-
-    }
-}
