+++ uml/scripts/comp.bat
