+++ uml/scripts/comp.sh
