+++ uml/src/main/java/com/unhandledexceptions/Model/FieldItem.java
@@ -1,4 +1,5 @@
-package Model;
+package com.unhandledexceptions.Model;
+
 public class FieldItem
 {
 	private String fieldName;
@@ -42,3 +43,4 @@ public class FieldItem
 		return this.type + " " + this.fieldName;
 	}
 }
+
